Upload the installation script and run it from the host.
the WinRM library base 64 encodes the powershell script and runs it inside a cmd using this one liner : "powershell -encodedCommand ...".
It has gotten too long and passes the 8191 character limit! until they come up with a better idea we can upload the script and then run it, or somehow run the script in a live session.